requirements.txt
# AI Fitness Mentor Bot — зависимости Python
# Установка: pip install -r requirements.txt
#
# Совместимость (проверено на PyPI, апрель 2026):
#   aiogram 3.27.0 требует pydantic>=2.4.1,<2.13
#   pydantic-settings 2.14.0 требует pydantic>=2.7.0
#   → pydantic 2.12.0 удовлетворяет оба ограничения

# ── Telegram Bot ──────────────────────────────────────────────────────────────
aiogram==3.27.0

# ── Конфигурация ──────────────────────────────────────────────────────────────
pydantic-settings==2.14.0
python-dotenv==1.0.1

# ── База данных ───────────────────────────────────────────────────────────────
supabase==2.29.0

# ── openai (используется как HTTP-клиент для Gemini OpenAI-совместимого API) ──
openai==2.32.0

# ── Планировщик задач ─────────────────────────────────────────────────────────
APScheduler==3.11.2

# ── HTTP-клиент (для Obsidian sync, скачивания фото)
# supabase 2.29.0 требует httpx>=0.26.0,<0.29
httpx==0.27.0

# ── Вспомогательные ───────────────────────────────────────────────────────────
# pydantic 2.12.0 — максимальная версия совместимая с aiogram 3.27.0 (<2.13)
pydantic==2.12.0
